Define LDBL_EPSILON, LDBL_MAX and LDBL_MIN as long double constants.
Submitted by:	Andreas Tobler <andreast-list@fgznet.ch>
Reviewed by:	das@


Former-commit-id: 235a063c75b40659096fddf583c633964ba53bfa

+++ sys/powerpc/include/float.h
@@ -286,7 +286,7 @@ begin_define
 define|#
 directive|define
 name|LDBL_EPSILON
-value|DBL_EPSILON
+value|(long double)DBL_EPSILON
 end_define
 
 begin_define
@@ -307,7 +307,7 @@ begin_define
 define|#
 directive|define
 name|LDBL_MIN
-value|DBL_MIN
+value|(long double)DBL_MIN
 end_define
 
 begin_define
@@ -328,7 +328,7 @@ begin_define
 define|#
 directive|define
 name|LDBL_MAX
-value|DBL_MAX
+value|(long double)DBL_MAX
 end_define
 
 begin_define
